removed the extra class and added logic in exporter
