+++ src/exporter.ts
@@ -6,6 +6,7 @@ import {
 } from "@credo-ts/askar/build/utils"
 import { writeFileSync } from "fs"
 import { join } from "path"
+import { Strategy } from "."
 
 /**
  * Class responsible for exporting wallet data.
@@ -14,6 +15,7 @@ export class Exporter {
   private walletConfig: WalletConfig
   private fileSystem: FileSystem
   private profileId: string
+  private tenantId: string
 
   /**
    * Constructor for the Exporter class.
@@ -33,12 +35,13 @@ export class Exporter {
     this.walletConfig = walletConfig
     this.fileSystem = fileSystem
     this.profileId = `tenant-${tenantId}`
+    this.tenantId = tenantId
   }
 
   /**
    * Export the wallet data to a JSON file.
    */
-  public async export(): Promise<void> {
+  public async export(strategy:string): Promise<void> {
     try {
       const askarWalletConfig = await this.getAskarWalletConfig(this.walletConfig)
       const store = await Store.open({
@@ -47,12 +50,17 @@ export class Exporter {
         passKey: askarWalletConfig.passKey,
       })
       console.log("🚀 Store opened:", store)
-      await this.getDecodedItemsAndTags(store)
+      if (strategy === Strategy.EXPORT) {
+        await this.getDecodedItemsAndTags(store);
+      } else if (strategy === Strategy.DB_PW_EXPORT) {
+        await this.processStoreData(store, askarWalletConfig);
+      } else {
+        console.warn(`Unknown strategy: ${strategy}`);
+      }
     } catch (error) {
       console.error("🚀 ~ Exporter ~ export ~ error:", error)
     }
   }
-
   /**
    * Get decoded items and tags from the store and write them to a JSON file.
    * @param store - The store instance.
@@ -82,6 +90,74 @@ export class Exporter {
     writeFileSync(outputPath, JSON.stringify(filteredData, null, 2))
     console.log(`Filtered data written to ${outputPath}`)
   }
+
+   private async processStoreData(store: any, askarWalletConfig): Promise<void> {
+      const data = await this.fetchDataFromStore(store);
+      const tenantRecord = this.findTenantRecord(data, this.tenantId);
+  
+      if (!tenantRecord) {
+        throw new Error('Tenant not found in the base wallet');
+      }
+  
+      const { walletId, walletKey } = this.extractWalletConfig(tenantRecord);
+      console.log(`Wallet ID: ${walletId}`);
+      console.log(`Wallet Key: ${walletKey}`);
+  
+      const baseUri = askarWalletConfig.uri.uri.split('/').slice(0, -1).join('/');
+      const tenantUri = `${baseUri}/${walletId}`;
+  
+      const tenantStore = await Store.open({
+        uri: tenantUri,
+        keyMethod: keyDerivationMethodToStoreKeyMethod(KeyDerivationMethod.Raw),
+        passKey: walletKey,
+      });
+  
+      const tenantData = await this.scanStore(tenantStore);
+      this.processTenantData(tenantData);
+    }
+  
+    private async fetchDataFromStore(store: any){
+      try {
+        const scan = store.scan({});
+        const records = await scan.fetchAll();
+        return records;
+      } catch (error) {
+        console.error("Error fetching data from store:", error);
+        throw new Error("Failed to fetch data from store");
+      }
+    }
+  
+    private findTenantRecord(data, tenantId: string) {
+      return data.find(record => record.category === 'TenantRecord' && record.name === tenantId);
+    }
+  
+    private extractWalletConfig(record): { walletId: string, walletKey: string } {
+      const value = JSON.parse(record.value);
+      return {
+        walletId: value.config.walletConfig.id,
+        walletKey: value.config.walletConfig.key,
+      };
+    }
+  
+    private async scanStore(store: any) {
+      const scan = store.scan({});
+      return await scan.fetchAll();
+    }
+  
+    private processTenantData(data): void {
+      const filteredData = {};
+  
+      for (const entry of data) {
+        if (!filteredData[entry.category]) {
+          filteredData[entry.category] = [];
+        }
+        filteredData[entry.category].push(entry);
+      }
+  
+      const outputPath = join(__dirname, "../DB_per_wallet_exportedWallet.json")
+      writeFileSync(outputPath, JSON.stringify(filteredData, null, 2))
+      console.log(`Filtered data written to ${outputPath}`)
+    }
 
   /**
    * Get the Askar wallet configuration.
